+++ keyboards/fallacy/rules.mk
@@ -24,7 +24,7 @@ AUDIO_ENABLE = no           # Audio output
 
 # project specific files
 SRC += indicators.c \
-       drivers/issi/is31fl3731-simple.c
+       drivers/led/issi/is31fl3731-simple.c
 QUANTUM_LIB_SRC += i2c_master.c
 
 LAYOUTS = alice alice_split_bs
