+++ keyboards/hs60/v2/ansi/rules.mk
@@ -31,6 +31,6 @@ LAYOUTS = 60_ansi
 # project specific files
 SRC =	keyboards/wilba_tech/wt_main.c \
 		keyboards/wilba_tech/wt_rgb_backlight.c \
-		drivers/issi/is31fl3733.c \
+		drivers/led/issi/is31fl3733.c \
 		quantum/color.c \
 		drivers/chibios/i2c_master.c
