+++ keyboards/nebula12/rules.mk
@@ -31,6 +31,6 @@ CIE1931_CURVE = yes
 # project specific files
 SRC +=  keyboards/wilba_tech/wt_main.c \
         keyboards/wilba_tech/wt_rgb_backlight.c \
-        drivers/issi/is31fl3731.c \
+        drivers/led/issi/is31fl3731.c \
         quantum/color.c
 QUANTUM_LIB_SRC += drivers/chibios/i2c_master.c
