+++ keyboards/nk87/nk87.c
@@ -18,7 +18,7 @@
 #endif
 
 #include "nk87.h"
-#include "drivers/issi/is31fl3733.h"
+#include "drivers/led/issi/is31fl3733.h"
 
 /* Indicator LEDS are part of the LED driver
  * Top LED is blue only. LED driver 2 RGB 63 Blue channel
