+++ keyboards/wilba_tech/rama_works_m60_a/rules.mk
@@ -44,5 +44,5 @@ LAYOUTS = 60_hhkb
 SRC =	keyboards/wilba_tech/wt_main.c \
 		keyboards/wilba_tech/wt_rgb_backlight.c \
 		quantum/color.c \
-		drivers/issi/is31fl3731.c \
+		drivers/led/issi/is31fl3731.c \
 		drivers/avr/i2c_master.c
