+++ keyboards/wilba_tech/rama_works_m6_b/rules.mk
@@ -41,5 +41,5 @@ CIE1931_CURVE = yes
 SRC =	keyboards/wilba_tech/wt_main.c \
 		keyboards/wilba_tech/wt_rgb_backlight.c \
 		quantum/color.c \
-		drivers/issi/is31fl3218.c \
+		drivers/led/issi/is31fl3218.c \
 		drivers/avr/i2c_master.c
