+++ keyboards/wilba_tech/wt60_b/rules.mk
@@ -42,5 +42,5 @@ CIE1931_CURVE = yes
 SRC =	keyboards/wilba_tech/wt_main.c \
 		keyboards/wilba_tech/wt_rgb_backlight.c \
 		quantum/color.c \
-		drivers/issi/is31fl3731.c \
+		drivers/led/issi/is31fl3731.c \
 		drivers/avr/i2c_master.c
