+++ keyboards/wilba_tech/wt80_a/rules.mk
@@ -30,7 +30,7 @@ BLUETOOTH_ENABLE = no       # Enable Bluetooth with the Adafruit EZ-Key HID
 AUDIO_ENABLE = no           # Audio output on port C6
 
 # project specific files
-SRC =	drivers/issi/is31fl3736.c \
+SRC =	drivers/led/issi/is31fl3736.c \
 		drivers/avr/i2c_master.c \
 		quantum/color.c \
 		keyboards/wilba_tech/wt_mono_backlight.c \
